+++ src/main/java/com/example/secumix/security/store/model/response/OrderDetailResponse.java
@@ -9,7 +9,7 @@ public class OrderDetailResponse {
     private int orderDetailId;
     private int quantity;
     private String productName;
-    private int priceTotal;
+    private long priceTotal;
     private String orderStatusName;
     private String address;
     private String productImg;
